Refactor, add commands folder
